requirements.txt
streamlit
pandas
google-generativeai
st-gsheets-connection